backend/requirements.txt
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
Django==6.0.2
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
pip==26.0.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
redis==7.2.1
sqlparse==0.5.5
daphne==4.1.0
